+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Metatable field `__debugtype` to set the displayed typename of an object
 * Show table keys with virtual children `<key>` and `<value>` and rename with unique names
 * Upload task now looks for zips in the correct place when Package is configured to place them outside the mod folder
+* Launch option `adjustModSettings` to update/clear mod settings before launching a debug session
 
 ## 1.1.1